+++ src/main/java/com/projeto/produto/dto/PermissionarioResponseDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PermissionarioResponseDTO {
-    private Long idPermissionario;
+    private Integer idPermissionario;
     private String numeroPermissao;
     private String nomePermissionario;
     private String cpfPermissionario;
